Abstract comment printing to a function
...called printComments in getShowComments.js
